refactor(frontend): extract PlaygroundSyncStateTag to separate component
